Character skills and update character_id table column type (#27)
* fix fitting relation issue and add a primary key on fitting_item table

* alter model in order to match with fitting item primary key

* implement skill endpoint
seed total sp and unallocated sp to character_info table

* add relation on character_skills
create a dedicated table for unallocated_sp and total_sp
fix character_id, corporation_id, alliance_id type according to latest CCP recommendation

* fix character_id type according to latest CCP recommendation

* implement agent research job [ok]

+++ src/database/migrations/2018_01_04_132252_create_character_fatigues_table.php
@@ -16,7 +16,7 @@ class CreateCharacterFatiguesTable extends Migration
 
         Schema::create('character_fatigues', function (Blueprint $table) {
 
-            $table->integer('character_id')->primary();
+            $table->bigInteger('character_id')->primary();
 
             $table->dateTime('last_jump_date')->nullable();
             $table->dateTime('jump_fatigue_expire_date')->nullable();
