+++ src/database/migrations/2018_01_05_130516_create_mail_bodies_table.php
@@ -16,7 +16,7 @@ class CreateMailBodiesTable extends Migration
 
         Schema::create('mail_bodies', function (Blueprint $table) {
 
-            $table->integer('mail_id')->primary();
+            $table->bigInteger('mail_id')->primary();
             $table->text('body');
 
             $table->timestamps();
